Move existing implementation to legacy.js
