+++ package.json
@@ -27,6 +27,8 @@
 	},
 	"dependencies": {
 		"@abcnews/alternating-case-to-object": "^2.0.1",
+		"@abcnews/env-utils": "^2.1.1",
+		"@abcnews/mount-utils": "^2.0.0",
 		"classnames": "^2.2.6",
 		"core-js": "^3.8.3",
 		"preact": "^10.5.12",
