+++ src/init.js
@@ -1,0 +1,7 @@
+const { h, render } = require('preact');
+const ExpandableCards = require('./components/ExpandableCards');
+const {selectMounts} = require('@abcnews/mount-utils');
+
+export const init = () => {
+  // TODO
+}
